Freelancer Rates | Discount finalized
It took way too long to realize the discount should be in the percentage so 13.5 should be / 100. The math happened to work out for the first 2 tests but the last one would error as the following:
```
* test apply_discount/2 it doesn't round (5.1ms) [L#33]

  1) test apply_discount/2 it doesn't round (FreelancerRatesTest)
     test/freelancer_rates_test.exs:33
     Expected the difference between 102.87962962962963 and 96.11015 (6.769479629629629) to be less than or equal to 1.0e-6
     code: assert_in_delta FreelancerRates.apply_discount(111.11, 13.5), 96.11015, 0.000001
     stacktrace:
       test/freelancer_rates_test.exs:34: (test)
```
I'm sure I don't need the percentage variable but I like to be verbose when it helps.