Pacman Rules | Win
Implement `win?` function using boolean `not` and `and` operators. I originally assumed this was an or operation which had passing tests but tests fail like
```
  * test win?/3 don't win if all dots eaten, but touching a ghost (2.4ms) [L#62]

  1) test win?/3 don't win if all dots eaten, but touching a ghost (RulesTest)
     test/rules_test.exs:62
     Expected false or nil, got true
     code: refute Rules.win?(true, false, true)
     stacktrace:
       test/rules_test.exs:63: (test)
```.
Without tests, I would've had to catch something like that in production. I also love this exercise because it's just enough reuse and conditional logic tends to be where one-off errors happen the most.