German sysadmin | Sanitize substitute final
Turns out we don't need `to_charlist`. The exercism site alerted me that the purpose of the exercise is the following:
```
The purpose of this exercise is to teach charlists. Solve it without using strings - do not use to_string, to_charlist, the String module, or the bitstring special form <<>>.
```