Roman numeral | Numeral up to 3000
Implement `numeral` 900 - 1000 and 1000+. This draws on a lot of what we covered with guards and recursion. I wouldn't have understood this implementation before covering the other exercises as I tried quite a while ago.

We go from bottom to top covering each guard clause. When we're 1 - 4, we prepend "I" then concat the results of the call - 1. If we start with say 3, the first call is I <> numeral(2), II <> numeral(1), III <> numeral(0). We take the same approach "going up the ladder" through each grouping as the digits work themselves out. Ranges form the same mental model but the special cases stick out more.