Library fees | Calculate the return date
Implement `return_date/1` to calculate the return date based on if the time is before noon or not. If it's before noon, add 28 days, otherwise it's 29. The add function also works in time increments only, so `:day` isn't a valid unit. It kinda seems like low hanging fruit to cover date changes like :day, :month, :year but maybe there's a reason not to?