Library fees | Determine how late the return of the book was
Implement `days_late/2` covering date differences. If the difference is greater than zero we clamp at zero. We care about the condition where the difference is < zero as we flip the negative number around. We approach this somewhat backwards as `Date.diff(actual_return_datetime, planned_return_date)` clamps at zero and doesn't seem to roll forward in the future, if I'm explaining this correctly.