Library fees | Determine if the book was returned on a Monday
Implement `monday?/1` to determine if the book was returned on Monday or 1 from the `day_of_week` function. Hints point to converting to a date but we don't need to do that, likely due to later Elixir revisions? It is more concise to change to a date to be more explicit but we're fine leaving this out.