Community Garden | 5. Get a registered plot
Implement the `CommunityGarden.get_registration/2` function. This seemed easy enough to stumble on though I checked my work. It makes sense that we would perform the same `list_registrations` call but try to find the value by plot_id in the list. Since our filter already gave us the pattern to search for, we reuse that and change `!=` to `==`.